+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Usage of search controller in frontend
 - Usage of scheduler generation in frontend
 
+### Removed
+
+- Unused fields in database
+
 ## 1.0.0
 
 ### Added